Modify new book feature editing the HTML, CSS, and JS
Edit HTML to markup the dialog of new book effectively. Replace parent container of dialog with div which will be displayed and hidden as required.
Change callback functions of form dialog open and form submit buttons to display and hide the modal.
Check for the form inputs after submission and before creation of new Book.
Use CSS class to display error of missing fields.
When all inputs are provided add a new book to library using addBookToLibrary() function.