Add Book read toggle and remove actions
Change id for all books in data.json to UUIDs so that they are refered to properly when calling the action functions.
In index.js declare libray HTML container selector globally.
Rename Book prototype function isReadTrue() to isReadToggle. The function toggles is read record on the object.
Modify book element header to display new is read toggle button and remove book button on an action bar.
Write toggleReadAction() and removeBookFromLibrary() to perform on click event for all actions, respectively.